Updatess to code
Handles greyscale by taking mean color
Creates new redscale grid image
Prints more useful information about sonification process

todo: Needs proper documentation
