--- tempDelete.txt
@@ -1,30 +1,0 @@
-Make the soniPics thing more OOP
-
-
-ToneClass:
-	Tone Lib
-	MinTones
-	MaxTones
-	ToneIncrements
-
-Image Class:
-	Image path
-	Image array
-	Image tones
-	Grid or Pixel
-	# For getting tone values:
-	Grid Maker
-	Pixel Maker
-	Thing for parsing image files - if we had flags to determine pixel or not
-	ToneClass object
-
-
-MidiMaker class
-	MidiMaker
-	Accepts ImageClass
-	# ImageClass will have
-	filename
-	tonelibrary
-
-	
-
